backend/requirements.txt
# requirements.txt
langchain
langchain-community
langchain-core
langchain-qdrant
langchain-openai
langchain
openai
python-dotenv
flask
flask-cors
gunicorn
tiktoken
resend

